pyproject.toml
[tool.poetry]
name = "instagram-emailer"
version = "0.1.0"
description = ""
authors = ["Bennett Yeagley <46436812+byeagley@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
instagrapi = "^2.1.2"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "ALL"
]